pyproject.toml
[tool.poetry]
name = "jpeg-to-pdf-converter"
version = "0.1.0"
description = "Flask API for jpeg to pdf conversion"
authors = ["ilkin-um <ilkinum@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
pytest = ">-7.1.2"
img2pdf = ">=0.4.4"
Pillow = ">=9.1.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
